:truck: Moving typescript tests to clients/typescript/bolt-sdk/test
