+++ Anchor.toml
@@ -39,7 +39,7 @@ members = [
 test = """
 echo "Waiting for 2 seconds..."
 sleep 2
-yarn run ts-mocha -p ./tsconfig.json -t 1000000 tests/typescript/main.ts
+yarn run ts-mocha -p ./tsconfig.json -t 1000000 clients/typescript/bolt-sdk/test/main.ts
 cd clients/csharp
 dotnet test
 """
