+++ clients/typescript/bolt-sdk/test/intermediate-level/permissioning/index.ts
