+++ clients/typescript/bolt-sdk/test/intermediate-level/permissioning/world.ts
@@ -5,7 +5,7 @@ import {
   ApplySystem,
   ApproveSystem,
   RemoveSystem,
-} from "../../../../clients/typescript/bolt-sdk/lib";
+} from "../../../lib";
 
 export function world(framework) {
   describe("World authority", () => {
