+++ clients/typescript/bolt-sdk/test/intermediate-level/world.ts
@@ -1,7 +1,7 @@
 import {
   InitializeNewWorld,
   InitializeRegistry,
-} from "../../../clients/typescript/bolt-sdk/lib";
+} from "../../lib";
 
 export function world(framework) {
   describe("World", () => {
